Fix tests and remove rememeber_config storage.
